+++ docs/components/layout/sidebar/Sidebar.tsx
@@ -1,9 +1,8 @@
-import Link from "next/link";
 import cl from "classnames";
 import { Close } from "@navikt/ds-icons";
 import style from "../layout.module.css";
 import { NextRouter } from "next/router";
-import Pages, { NavdsPage } from "../../../src/pages";
+import Menu from "./Menu";
 
 interface SidebarProps {
   classname?: string;
@@ -14,53 +13,6 @@ interface SidebarProps {
 }
 
 const Sidebar = ({ sidebar, small, onSidebarChange, route }: SidebarProps) => {
-  const subLinks = (page: NavdsPage) => {
-    return (
-      <div key={page.title}>
-        <span>{page.title}</span>
-        <ul id={"collapse " + page.title}>
-          {page.children.map((child) => {
-            if (child.pathName) {
-              return (
-                <li key={child.pathName}>
-                  <Link href={child.pathName}>
-                    {child.title ? child.title : parseTitle(child.pathName)}
-                  </Link>
-                </li>
-              );
-            } else if (child.children && child.title) {
-              return subLinks(child);
-            }
-            return null;
-          })}
-        </ul>
-      </div>
-    );
-  };
-
-  const parseTitle = (path: string) =>
-    path
-      .split("/")
-      .pop()
-      .replace(/^\w/, (c) => c.toUpperCase())
-      .replace("-", " ") || "Default Title";
-
-  const generateLinks = () => {
-    return Pages.map((page) => {
-      if (page.pathName) {
-        return (
-          <Link key={page.pathName} href={page.pathName}>
-            {page.title ? page.title : parseTitle(page.pathName)}
-          </Link>
-        );
-      } else if (page.title && page.children) {
-        return subLinks(page);
-      } else {
-        return null;
-      }
-    });
-  };
-
   return (
     <>
       {sidebar && small && (
@@ -85,10 +37,7 @@ const Sidebar = ({ sidebar, small, onSidebarChange, route }: SidebarProps) => {
             <Close />
           </button>
         )}
-        {generateLinks()}
-        {/* <Link href="/">Home</Link>
-        <Link href="/komponenter">Komponenter</Link>
-        <Link href="/komponenter/button">Button</Link> */}
+        <Menu />
       </div>
     </>
   );
